+++ src/types/ReviewFinder.ts
@@ -1,4 +1,6 @@
+import { Dispatch, AnyAction } from 'redux';
 import { MultiValue } from 'react-select';
+import { RootState } from '../store/store';
 
 /* eslint-disable */
 
@@ -9,6 +11,8 @@ export type reviewBlackList = MultiValue<reviewerType>;
 export type showReviewerType = {'id': number, login: string, avatar: string};
 
 export const showReviewerDefault: showReviewerType = {id: 0, login: '', avatar: ''};
-export const reviewersDefault: reviewerType ={'value': 0, 'label': 'Логины не загружены', 'avatar': '', 'isDisabled': true}
+export const reviewersDefault: reviewerType ={'value': 0, 'label': 'Ревьюеры не загружены', 'avatar': '', 'isDisabled': true}
 
-
+export type FetchUserDataFromGithub = ( dispatch: Dispatch, getState: () => RootState ) => Promise<void>;
+export type DispatchSettings = (arg: FetchUserDataFromGithub | AnyAction) => Promise<void>
+export type FetchUserData = () => FetchUserDataFromGithub;
